requirements.txt
scapy
flask
requests
colorama
rich
python-dotenv
pyyaml
scikit-learn
numpy